// apps/ripscloud-web/src/api/api.ts
export interface RipsAdminApplicationDTOsCreateSpecialistRequestDto {
    'userId'?: string;
    'documentType'?: string;
    'documentNumber'?: string;
    'professionalType'?: string;
    'registrationNumber'?: string;
    'birthDate'?: string;
    'sexCode'?: string;
    'countryResidenceCode'?: string;
    'municipalityResidenceCode'?: string;
    'territorialZoneCode'?: string;
    'isActive'?: boolean;
}
// ...
export interface RipsAdminApplicationDTOsSpecialistDto {
    'id'?: string;
    'tenantId'?: string;
    'user'?: RipsAdminApplicationDTOsSpecialistUserDto;
    'documentType'?: string;
    'documentNumber'?: string;
    'professionalType'?: string;
    'registrationNumber'?: string;
    'birthDate'?: string;
    'sexCode'?: string;
    'countryResidenceCode'?: string;
    'municipalityResidenceCode'?: string;
    'territorialZoneCode'?: string;
    'isActive'?: boolean;
    'createdAt'?: string;
    'updatedAt'?: string | null;
}
export interface RipsAdminApplicationDTOsSpecialistUserDto {
    'userId'?: string;
    'email'?: string;
    'firstName'?: string | null;
    'lastName'?: string | null;
}
// ...
export interface RipsAdminApplicationDTOsUpdateSpecialistRequestDto {
    'documentType'?: string;
    'documentNumber'?: string;
    'professionalType'?: string;
    'registrationNumber'?: string;
    'birthDate'?: string;
    'sexCode'?: string;
    'countryResidenceCode'?: string;
    'municipalityResidenceCode'?: string;
    'territorialZoneCode'?: string;
    'isActive'?: boolean;
}
// ...
export interface RipsAdminApplicationDTOsWorkspaceUserDto {
    'userId'?: string;
    'email'?: string;
    'firstName'?: string;
    'lastName'?: string;
    'roleName'?: string;
    'isActive'?: boolean;
    'joinedAt'?: string;
}